requirements.txt
asgiref==3.6.0
certifi==2020.6.20
dj-database-url==0.5.0
Django==3.2
django-crispy-forms==1.9.2
gunicorn==20.0.4
pafy==0.5.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
youtube-dl==2020.12.14
djangorestframework==3.14.0
django_cors_headers==4.1.0
pytube==15.0.0
instaloader==4.9.6
db-sqlite3==0.0.1
graphene-django==3.1.2
firebase_admin==6.2.0
apify_client==1.2.2